requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-dateutil>=2.8.0
xlsxwriter>=3.1.0
chardet>=5.0.0
